requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
cssselect2==0.7.0
Django==4.2.16
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-simple-history==3.7.0
django_csp==3.8
et_xmlfile==2.0.0
fonttools==4.55.3
openpyxl==3.1.5
pillow==11.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.0
python-dotenv==1.0.1
pytz==2024.2
reportlab==4.2.5
sentry-sdk==2.19.2
sqlparse==0.5.2
tablib==3.7.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
weasyprint==63.1
webencodings==0.5.1
XlsxWriter==3.2.0
xlwt==1.3.0
zopfli==0.2.3.post1
